splitting off fco metrics testing to separate file and updating tests for re-factored electricity generated metric
